fix(circular-deps): prevent circular dependencies in components
