+++ src/components/ConnectionConfig.styles.ts
@@ -1,0 +1,5 @@
+import { css } from '@emotion/css';
+
+export const assumeRoleInstructionsStyle = css({
+  maxWidth: '715px',
+});
