+++ src/components/SIGV4ConnectionConfig.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import { DataSourcePluginOptionsEditorProps, DataSourceSettings } from '@grafana/data';
-import { ConnectionConfig, ConnectionConfigProps } from '../components/ConnectionConfig';
+import { ConnectionConfig } from '../components/ConnectionConfig';
 
-import { AwsAuthDataSourceSecureJsonData, AwsAuthDataSourceJsonData } from '../types';
+import { AwsAuthDataSourceSecureJsonData, AwsAuthDataSourceJsonData, ConnectionConfigProps } from '../types';
 
 export interface SIGV4ConnectionConfigProps extends DataSourcePluginOptionsEditorProps<any, any> {
   inExperimentalAuthComponent?: boolean;
